+++ app/src/main/java/com/safarparmar/app/domain/model/HomeModels.kt
@@ -135,8 +135,8 @@ data class Achievement(
     val holderCount: Int = 0,
     val earned: Boolean = false,
     val progress: Int = 0,
-    val currentValue: Int = 0,
-    val targetValue: Int = 0
+    val currentValue: Double = 0.0,
+    val targetValue: Double = 0.0
 )
 
 @Immutable
